Fix : properties 문서 gitignore에 추가 및 변경사항 다시 재커밋

+++ src/main/java/com/sanaiclub/contract/model/vo/ContractMilestoneVO.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 @Builder
 public class ContractMilestoneVO {
     
-    private final Long milestoneId;
+    private final Integer milestoneId;
     private final Integer contractId;
     private final Integer step;
     private final String title;
